Changed rr_indexed_data from array to vector

+++ vpr/src/route/rr_node.cpp
@@ -51,7 +51,7 @@ short t_rr_node::class_num() const {
     return ptc_.class_num;
 }
 
-short t_rr_node::cost_index() const {
+size_t t_rr_node::cost_index() const {
 	return cost_index_;
 }
 
@@ -173,7 +173,7 @@ void t_rr_node::set_class_num(short new_class_num) {
     ptc_.class_num = new_class_num;
 }
 
-void t_rr_node::set_cost_index(short new_cost_index) {
+void t_rr_node::set_cost_index(size_t new_cost_index) {
 	cost_index_ = new_cost_index;
 }
 void t_rr_node::set_rc_index(short new_rc_index) {
